Add check license, complete installer, refactoring

+++ ЗИ ЛР1/FileCoding.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using Microsoft.Win32;
+using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -13,22 +15,43 @@ namespace ЗИ_ЛР1
 	{
 		public FileCoding(Form form)
 		{
-			ConfirmPasswordForm confirmPasswordForm = new ConfirmPasswordForm("Введите контрольную фразу");
-			if (confirmPasswordForm.ShowDialog() == DialogResult.OK)
+			try
 			{
-				try
+				RSAParameters publicKey = JsonConvert.DeserializeObject<RSAParameters>(File.ReadAllText("openKey.json"));
+				RSACryptoServiceProvider rSACryptoServiceProvider = new RSACryptoServiceProvider();
+				rSACryptoServiceProvider.ImportParameters(publicKey);
+				byte[] signedData = (byte[])Registry.CurrentUser.OpenSubKey("Software").OpenSubKey("Statnikov A.S.").GetValue("Signature");
+				byte[] systemInfo = Encoding.ASCII.GetBytes(new SystemInfo().getTotalInfo());
+				if (rSACryptoServiceProvider.VerifyData(systemInfo, CryptoConfig.MapNameToOID(HashAlgorithmName.MD5.Name), signedData))
 				{
-					DecodeFile("usersCoded", "users", MD4(Encoding.ASCII.GetBytes(confirmPasswordForm.password)));
-					File.Delete("usersCoded");
-					form.ShowDialog();
-					CodeFile("users", "usersCoded", MD4(Encoding.ASCII.GetBytes(confirmPasswordForm.password)));
-					File.Delete("users");
+					ConfirmPasswordForm confirmPasswordForm = new ConfirmPasswordForm("Введите контрольную фразу");
+					if (confirmPasswordForm.ShowDialog() == DialogResult.OK)
+					{
+						try
+						{
+							DecodeFile("usersCoded", "users", MD4(Encoding.ASCII.GetBytes(confirmPasswordForm.password)));
+							File.Delete("usersCoded");
+							form.ShowDialog();
+							CodeFile("users", "usersCoded", MD4(Encoding.ASCII.GetBytes(confirmPasswordForm.password)));
+							File.Delete("users");
+						}
+						catch
+						{
+							MessageBox.Show("Ошибка контрольной фразы");
+							Application.Exit();
+						}
+					}
 				}
-				catch
+				else
 				{
-					MessageBox.Show("Ошибка контрольной фразы");
+					MessageBox.Show("Ошибка: несанкционировнное копирование");
 					Application.Exit();
 				}
+			}
+			catch
+			{
+				MessageBox.Show("Ошибка: несанкционировнное копирование");
+				Application.Exit();
 			}
 		}
